:sparkles: 리뷰 미디어 repository 분리 및 삭제시 media 삭제
