+++ src/main/kotlin/sparta/nbcamp/wachu/domain/review/repository/v1/ReviewMultiMediaJpaRepository.kt
@@ -5,4 +5,5 @@ import sparta.nbcamp.wachu.domain.review.model.v1.ReviewMultiMedia
 
 interface ReviewMultiMediaJpaRepository : JpaRepository<ReviewMultiMedia, Long> {
     fun findAllByReviewId(reviewId: Long): List<ReviewMultiMedia>
+    fun deleteAllByReviewId(reviewId: Long)
 }
