+++ src/main/kotlin/sparta/nbcamp/wachu/domain/review/service/v1/ReviewServiceImpl.kt
@@ -11,6 +11,7 @@ import sparta.nbcamp.wachu.domain.review.dto.v1.ReviewRequest
 import sparta.nbcamp.wachu.domain.review.dto.v1.ReviewResponse
 import sparta.nbcamp.wachu.domain.review.model.v1.ReviewMediaType
 import sparta.nbcamp.wachu.domain.review.model.v1.ReviewMultiMedia
+import sparta.nbcamp.wachu.domain.review.repository.v1.ReviewMultiMediaRepository
 import sparta.nbcamp.wachu.domain.review.repository.v1.ReviewRepository
 import sparta.nbcamp.wachu.domain.wine.repository.WineRepository
 import sparta.nbcamp.wachu.exception.AccessDeniedException
@@ -24,6 +25,7 @@ class ReviewServiceImpl(
     private val wineRepository: WineRepository,
     private val memberRepository: MemberRepository,
     private val reviewRepository: ReviewRepository,
+    private val reviewMediaRepository: ReviewMultiMediaRepository,
     private val mediaS3Service: MediaS3Service
 ) : ReviewService {
     override fun getReviewPage(pageable: Pageable): Page<ReviewResponse> {
@@ -35,7 +37,7 @@ class ReviewServiceImpl(
         val review = reviewRepository.findById(id)
             ?: throw ModelNotFoundException("Review", id)
 
-        val mediaList = reviewRepository.mediaFindAll(id).map { ReviewMultiMediaResponse.from(it) }
+        val mediaList = reviewMediaRepository.mediaFindAll(id).map { ReviewMultiMediaResponse.from(it) }
 
         return ReviewResponse.from(review, mediaList)
     }
@@ -57,13 +59,14 @@ class ReviewServiceImpl(
         if (!images.isNullOrEmpty()) {
             mediaList = mediaS3Service.upload(images, S3FilePath.REVIEW.path + "${review.id}/")
                 .let { it.map { url -> ReviewMultiMedia.toEntity(review.id!!, url, ReviewMediaType.IMAGE) } }
-                .let { reviewRepository.mediaSave(it) }
+                .let { reviewMediaRepository.mediaSave(it) }
                 .let { it.map { multiMedia -> ReviewMultiMediaResponse.from(multiMedia) } }
         }
 
         return ReviewResponse.from(review, mediaList)
     }
 
+    @Transactional
     override fun deleteReview(userPrincipal: UserPrincipal, id: Long) {
         val review = reviewRepository.findById(id)
             ?: throw ModelNotFoundException("Review", id)
@@ -74,6 +77,7 @@ class ReviewServiceImpl(
             )
         ) { throw AccessDeniedException("not your review") }
 
+        reviewMediaRepository.deleteAllByReviewId(id)
         reviewRepository.delete(review)
     }
 
@@ -94,10 +98,10 @@ class ReviewServiceImpl(
 
         val mediaList = mediaS3Service.upload(multipartFileList, S3FilePath.REVIEW.path + "$reviewId/")
             .let { it.map { url -> ReviewMultiMedia.toEntity(reviewId, url, ReviewMediaType.IMAGE) } }
-        return reviewRepository.mediaSave(mediaList).map { ReviewMultiMediaResponse.from(it) }
+        return reviewMediaRepository.mediaSave(mediaList).map { ReviewMultiMediaResponse.from(it) }
     }
 
     override fun getReviewMultiMedia(reviewId: Long): List<ReviewMultiMediaResponse> {
-        return reviewRepository.mediaFindAll(reviewId).map { ReviewMultiMediaResponse.from(it) }
+        return reviewMediaRepository.mediaFindAll(reviewId).map { ReviewMultiMediaResponse.from(it) }
     }
 }
